[Repl] Support Initial Commands, &&, print Dargon logo.
Admittedly the Intro for DMI shouldn't be in the REPL code.
But I want it printed when DMI is embedded in other apps.

Perhaps it'd be better to create a DmiReplCore. Not sure, that also sounds
gross. My current approach is to extend Dispatcher with DMI commands, so
unless ReplCore looks at dispatcher for the intro (wtf?) I'm not sure how
to even get an intro from DMI to that cleanly... unless perhaps we're
calling the DMI logo print in this "add DMI commands" function... which
functionally works but would also be pretty wtf.

Initial commands are, for example "use tcp && fetch-mobs && tree", which
is separately executed one at a time (stopping execution of the following
if one fails).