[Commons] Add EnumeratorToEnumerableAdapter
C#'s foreach relies on a duck-typed ienumerable which implements method
GetEnumerator(). This allows types like List<T> to return a struct from
GetEnumerator() to avoid a GC alloc.

This return value (even if a struct) must implement IEnumerator<T>. The
enumerator object is akin to an iterator, while the IEnumerable-like
(with GetEnumerator) is akin to a collection view.

A read-only collection view can be represented by the initial state of an
iterator. C# provides no such adapter for this. To serve this purpose,
I've written this helper class which exposes an iterator as a
IEnumerable-like collection view.