[Commons] improvements to Dfs/Bfs variants
DFS throws pushes into a temporary stack, then pops them in reverse
order. This is necessary to get a correct traversal (not reversed within
a level) because the backing structure is a queue.

Both DFS and BFS avoid allocating a hashset & running seen / clear
operations if the operation has deduping disabled.