[ryu] support container naming, asynchronous container initialization with Eventual, and RyuType merging with DenyDefaultActivate.
Container Naming makes it easy to tell what a RyuContainer does at a glance.

As initialization is now asynchronous, support for tuple-declared singletons
works via adding a node for the tuple (which is computed potentially on main thread)
and then awaiting that result from individually exposed tuple items. Furthermore, it's
no longer sensical straightforward to support a second-phase Initialize, which was
broken and unused prior anyway, so this has been removed.

The new system also supporst requesting that a type's activator be executed on the
main synchronization context, which is useful for windowing scenarios. Prior, explicitly
declaring this wasn't necessary as all container initialization ran on main thread. Conceptually,
every instantiation is now a task, and dependencies work by awaiting other tasks.

The main & background synchronization contexts are specified via the
RyuSynchronizationContexts type.