[vox] support serializing non-auto properties and properties in base class of a different assembly
Source generators do not have access to the private metadata of assemblies they depend upon. For that reason,
the existing code failed to serialize auto-properties in the base class of a type, if the base class was in a different
assembly (as then its private field wouldn't be visible). As a workaround, serialize all properties with a getter/setter
except for the ones which are indexers, and serialize fields which aren't implicitly declared / const.